+++ HealthStaminaRegen/HealthStaminaRegen/Config/ModConfig.cs
@@ -11,5 +11,6 @@ namespace HealthStaminaRegen.Config
     {
         public HealthConfig Health { get; set; } = new HealthConfig();
         public StaminaConfig Stamina { get; set; } = new StaminaConfig();
+        public bool IgnorePause { get; set; } = false;
     }
 }